Added custom middleware to add a user to the req express object. The user is then passed into context through gqlExpress.
The addUser middleware checks if there is a valid token, then tries to verify its authenticity with the token and secret. If its good, req.user is set to user and next() is called. If there is an error here, then the token is invalid (expired or bad secret), so we try to refresh the token. If this works, new response headers are sent back to the client 'x-token' and 'x-refresh-token'. Then req.user is set to user and next() called