Added the newChannelMessage subscription to the Messages Schema and created a resolver for it
The resolver listens for a 'NEW_CHANNEL_MESSAGE' pubsub event, then checks if the passed argument message's channelId matches the target payload's channelId. Now inside the createMessage mutation, we publish a 'NEW_CHANNEL_MESSAGE' pubsub event containing sequelize dataValues for the message, and the currentUser. Although I am resolving the 'user' field for all message gql queries, it is not being found in the subscription so we find it manually here inside an async function as a work aorund