Added DirectMessage model to sequelize and created its gql Schema
DirectMessage's have a 'text' field for the message contents, and several foreign key fields. They belong to User models w/ two foriegn key as user_id's for both the sender and receiver. DirectMessages also belong to a Team, because only users on the same team may send each other direct messages (and as such these messages reside in the context of a team)