Configured apollo-server-upload to accept files and store them in a directory named files/
Set up some required stuff for the library to work correctly - the schema needs  and the type File. file arguments for mutations are of type Upload. The resolver needs Upload: GraphQLUpload (imported from apollo-server-upload). Now a message will have either text, or a file. If there is a file present in the mutation resolver, we will attempt to first store it in the filesystem at /files/, then we will attach the url and mimetype to the message object and store those in the DB (message model has been updated to reflect these two new fields)