New simple model (table) to store members of private channels
The table stores channelId and userId of pcmember records. For the resolver of createChannelMutation, if public is passed to args as false, then we will map over the supplied members list, and set the userId and channelId for each pcmember before they are stored into the PCMember model. We use sequelize's bulkCreate to store all of the pcmembers at once, and this operation is a part of a transaction with the channel creation itself, so that if the channel is private and we need to add pcmembers, the channel won't get created unless that operation succeeds