Wrote the getOrCreateChannel mutation
This mutation will first check if a dm message group already exists between the selected users, if so the mutation will return (get) that channel. If one doesn't exist in the DB, we will create a new channel for it. The idea is to reuse channels for public channels and messages between users