Switched from offset to cursor pagination strategy
The cursor is the 'created_at' String date of a message. If a cursor is present in the messages query, the query will resolve back the specified limit of messages after that cursor date. NOTE: the cursor and the sorted by options shoould be the same (created_at). For optimization, we also make the created_at field an index inside the message model.